+++ lib/collaboration_web/controllers/user_controller.ex
@@ -1,32 +1,44 @@
 defmodule CollaborationWeb.UserController do
   use CollaborationWeb, :controller
 
-  import Collaboration.Accounts
-
+  alias Collaboration.Accounts
   alias Collaboration.Accounts.User
 
-  # start page to create experiment users
+  @doc """
+  Allows to start an experiment by providing a form to create a participant.
+  Serves as the landing page for unauthenticated users.
+  """
   def new(conn, _params) do
-    render conn, "new.html", changeset: User.changeset(%User{})
+    render conn, "new.html", changeset: Accounts.change_user(%User{})
   end
 
+  @doc """
+  Create a participant user. On success start experiment by showing topic.
+  """
   def create(conn, %{"user" => user_params }) do
-    case create_user(user_params) do
+    case Accounts.create_user(user_params) do
       {:ok, user} ->
         conn
         |> CollaborationWeb.Auth.login(user)
         |> redirect(to: Routes.topic_path(conn, :show))
 
-      {:error, changeset} ->
+      {:error, %Ecto.Changeset{} = changeset} ->
+        IO.inspect changeset
         render conn, "new.html", changeset: changeset
     end
   end
 
+  @doc """
+  Shows a list of bot and admin users.
+  """
   def index(conn, _) do
-    render conn, "index.html", users: list_users()
+    render conn, "index.html", users: Accounts.list_users()
   end
 
+  @doc """
+  Shows a list of participant users.
+  """
   def participants(conn, _) do
-    render conn, "participants.html", users: list_participants()
+    render conn, "participants.html", users: Accounts.list_participants()
   end
 end